requirements.txt
Django==5.2.7
python-dotenv==1.1.1
dj-database-url
psycopg2-binary==2.9.5
gunicorn
whitenoise
django-encrypted-model-fields==0.6.5
pillow==11.3.0
moviepy==2.2.1
openai==2.3.0
requests==2.32.3
beautifulsoup4==4.12.3
openpyxl==3.1.4
pandas==2.2.2